feat: allow use of external or internal framebuffer
To avoid copying data unneccesarily, we can choose to use an internal representation for the frame buffer, which is useful for WASM, since passing a long lived mutable reference across the ABI is something I have been unable to implement.
Allowing access to an external frame buffer is more suited for a complete Rust solution, however we have to use unsafe code to create a mutable slice that avoid the use of lifetimes. This avoidance is again due to compatibility with WASM, specifically wasm_bindgen as we cannot currently export structs with lifetime annotations.